requirements.txt
streamlit
python-dotenv
PyPDF2
google-generativeai

langchain  # core framework
langchain-huggingface #connect huggingface models to perform embedding
faiss-cpu #fast  vector db to store embedded data
langchain-community #extra integration
langchain-text-splitters #to split large data into smaller chunks
sentence-transformers #pretrained embedding models to convert text to vectors
langchain-core #document,chain,etc..
